+++ Sources/SurfaceRecorderSDK/ARViewFrameSource.swift
@@ -1,6 +1,6 @@
 //
 //  ARViewFrameSource.swift
-//  ReelKit
+//  SurfaceRecorderSDK
 //
 //  Created by Nazar Kozak on 05.06.2026.
 //