+++ Sources/SurfaceRecorderSDK/FrameConverter.swift
@@ -1,6 +1,6 @@
 //
 //  FrameConverter.swift
-//  ReelKit
+//  SurfaceRecorderSDK
 //
 //  Created by Nazar Kozak on 05.06.2026.
 //