+++ Sources/SurfaceRecorderSDK/FrameSource.swift
@@ -1,6 +1,6 @@
 //
 //  FrameSource.swift
-//  ReelKit
+//  SurfaceRecorderSDK
 //
 //  Created by Nazar Kozak on 05.06.2026.
 //