+++ Sources/SurfaceRecorderSDK/UIViewFrameSource.swift
@@ -1,6 +1,6 @@
 //
 //  UIViewFrameSource.swift
-//  ReelKit
+//  SurfaceRecorderSDK
 //
 //  Created by Nazar Kozak on 05.06.2026.
 //